# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Huynh Land Firebase API
  version: 1.0.0
paths:
...
  /api/mission:
    get:
      tags: [Mission]
      summary: Hiển thị sứ mệnh theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Lấy sứ mệnh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  mission:
                    type: object
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có sứ mệnh
    post:
      tags: [Mission]
      summary: Tạo sứ mệnh theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Sứ mệnh đã được tạo
        "400":
          description: Sứ mệnh đã tồn tại hoặc ngôn ngữ không hợp lệ
    put:
      tags: [Mission]
      summary: Sửa sứ mệnh theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "200":
          description: Sứ mệnh đã được cập nhật
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có sứ mệnh để cập nhật
    delete:
      tags: [Mission]
      summary: Xóa sứ mệnh theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Sứ mệnh đã được xóa
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có sứ mệnh để xóa
  /api/vision:
    get:
      tags: [Vision]
      summary: Hiển thị tầm nhìn theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Lấy tầm nhìn thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  vision:
                    type: object
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có tầm nhìn
    post:
      tags: [Vision]
      summary: Tạo tầm nhìn theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Tầm nhìn đã được tạo
        "400":
          description: Tầm nhìn đã tồn tại hoặc ngôn ngữ không hợp lệ
    put:
      tags: [Vision]
      summary: Sửa tầm nhìn theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "200":
          description: Tầm nhìn đã được cập nhật
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có tầm nhìn để cập nhật
    delete:
      tags: [Vision]
      summary: Xóa tầm nhìn theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Tầm nhìn đã được xóa
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có tầm nhìn để xóa
...
  /api/office:
    get:
      tags: [Office]
      summary: Hiển thị địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Lấy địa chỉ văn phòng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  office:
                    type: object
                    properties:
                      name:
                        type: string
                      address:
                        type: string
                      phone:
                        type: string
                      gmail:
                        type: string
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có địa chỉ văn phòng
    post:
      tags: [Office]
      summary: Tạo địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                phone:
                  type: string
                gmail:
                  type: string
      responses:
        "201":
          description: Đã tạo địa chỉ văn phòng
        "400":
          description: Địa chỉ văn phòng đã tồn tại hoặc thiếu trường
    put:
      tags: [Office]
      summary: Sửa địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                phone:
                  type: string
                gmail:
                  type: string
      responses:
        "200":
          description: Đã cập nhật địa chỉ văn phòng
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có địa chỉ văn phòng để cập nhật
    delete:
      tags: [Office]
      summary: Xóa địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Đã xóa địa chỉ văn phòng
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có địa chỉ văn phòng để xóa
  /api/social:
    get:
      tags: [Social]
      summary: Lấy thông tin kết nối với chúng tôi
      responses:
        "200":
          description: Lấy thông tin social thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  social:
                    type: object
                    properties:
                      facebook:
                        type: string
                      youtube:
                        type: string
        "404":
          description: Không có thông tin social
    post:
      tags: [Social]
      summary: Tạo thông tin kết nối với chúng tôi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                facebook:
                  type: string
                youtube:
                  type: string
      responses:
        "201":
          description: Đã tạo thông tin social
        "400":
          description: Thông tin social đã tồn tại hoặc thiếu trường
    put:
      tags: [Social]
      summary: Sửa thông tin kết nối với chúng tôi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                facebook:
                  type: string
                youtube:
                  type: string
      responses:
        "200":
          description: Đã cập nhật thông tin social
        "404":
          description: Không có thông tin social để cập nhật
    delete:
      tags: [Social]
      summary: Xóa thông tin kết nối với chúng tôi
      responses:
        "200":
          description: Đã xóa thông tin social
        "404":
          description: Không có thông tin social để xóa
...